refactor: extract mapFileToStatementFormat method to make it easier to test
